+++ src/index.html
@@ -8,22 +8,22 @@
       crossorigin="anonymous" referrerpolicy="no-referrer">
     <title>ToDo List</title>
   </head>
-  <body class="bg-container">
+  <body class="backgd-container">
     <div class="toDo-list">
       <h2>ToDo List</h2>
-      <div class="intro">
+      <div class="Intro">
         <h3>Today's ToDo</h3>
         <i class="fa-sharp fa-solid fa-arrows-rotate"></i>
       </div>
       <hr class="line-1">
       <div class="input-data">
-        <input class="inputs-field" id="input" type="text" placeholder="Add to your list...">
+        
+        <input class="inputs-field" id="input" type="text" placeholder="Add to your list.." required> 
         <button class="add-btn" type="submit">+</button>
       </div>
       <hr class="line-1">
       <div class="list"></div>
       <button id="clear-btn">Clear all completed</button>
-      </div>
-      <script type="module" src="./index.js"></script>
+    </div>
   </body>
-  </html>
+</html>
